Moving to .net 9 and adding a bit of server code for fse widget

+++ CommonHelpers/Vlv0100.cs
@@ -208,7 +208,7 @@
         }
         private void SetRampRate(byte rampRate)
         {
-            byte[] data = BitConverter.GetBytes(rampRate);
+            byte[] data = BitConverter.GetBytes((short)rampRate);
             inpOut?.WriteMemory(FRPR, data);
         }
     }
